Add cargo-make dependency and makefiles
Cargo-make is a task runner for cargo projects that allows defining build tasks and managing task dependencies. For engineers with Java background, cargo-make analog is Ant.

All tasks are defined and configured via toml files. This change adds `Makefile.toml` that contains definitions of most commonly used tasks and will be extended in the future and `grcov.makefile.toml` that contains the definition of code coverage task.

Alternatives include using pure make that is powerful enough to suit the most needs, however cargo-make is slightly easier to start with for a Rust developer.

Some examples of how to use it:

## Installation

```
cargo install cargo-make
```

## Run tasks

Run check, format and lints

```bash
  cargo make touch
```

Run touch + tests and doc tests

```bash
  cargo make san
```