Migrate to `tracing` crate
Tracing is a superset of features we're using currently (print to stderr, verbosity, format) and is required to instrument MPC helpers with telemetry required to understand the performance of it and different protocols we'll be experimenting with.

This change also adds some pretty basic tracing of HTTP requests by using `tower-http` crate.

`stderrlog` crate is removed as we no longer need to use it. Besides that, it has a security advisory against it

```
Crate:     time
Version:   0.1.44
Title:     Potential segfault in the time crate
Date:      2020-11-18
ID:        RUSTSEC-2020-0071
URL:       https://rustsec.org/advisories/RUSTSEC-2020-0071
Solution:  Upgrade to >=0.2.23
Dependency tree:
time 0.1.44
└── chrono 0.4.19
    └── stderrlog 0.5.3
        └── raw-ipa 0.1.0

error: 1 vulnerability found!
```