Use the MMO^pi function for PRSS
I'm not entirely clear on the relative benefits of using this over the
fixed-key model of AES - particularly as we are reducing the `u128`
modulo $p << 2^128$ - but this seems like a relatively cheap way to
build the function.  This function provides correlation robustness,
whereas the fixed-key mode used previously does not.