Add Helper Identity enum
There seem to be more and more protocols we need to implement that require helpers to know their exact position inside the MPC ring. For instance, modulus conversion needs to construct binary secret shares differently depending on which helper performs this operation (#63).

Secure sort (#66) seem to need that as well as it requires helpers to agree on constructing a secret share of `1`.

This change adds necessary plumbing to make helper identity available for protocol writers - `ctx.helper_ring.identity()` should return the exact helper location