Add ProtocolStep enum
As discussed offline with @martinthomson and @thurstonsand, we need to model protocol step as an enum.

Effectively, in the long term we will have a composite id that uniquely identifies every multiplication happening in the system that consists of:

* Query id (not in this PR). Every query has its own circuit created for their own purposes.
* Record id (not in this PR) - monotonically increasing integer that uniquely identifies every record (secret share) in the input.
* Step (this PR). Unique transformation on a given record. Effectively used to uniquely identify multiply operation as it is the only one currently that requires communication.

Steps form a tree where only leaf nodes are used in the communication and intermediate nodes are classifiers.

A combination of these three identifiers allows each helper to correctly interpret every value sent towards it.

Right now our mental model about communication includes every helper to create a new request per query + step (leaf node inside `ProtocolStep` enum):

```
https://<helper_endpoint>/api/query/<query_id>/<step>

[values_stream]
```

record id is the index in the values array.