Remove spin and wait from receive
Previously `receive` was running a tight loop where on each iteration it checked whether the message is available or not. This change makes all the hard work happening inside one task that holds onto a local buffer of messages, so we don't need any synchronization there. The communication between the event loop and `receive` function is done via oneshot channels