Move send buffers from network to infra layer
As discussed offline with Martin, we want to keep the network layer simple.
That means that all the buffering logic on send and receive should be handled
by the infrastructure layer. Receive buffers already exist there, this change
lifts the send buffers to it.

Right now the implementation is dumb and just accepts whatever the protocols send
to it. There is no accepted record id interval or any other mechanism to block
the protocol if it advanced too far. That will be the next step.

The main purpose of this change is to get us to agree on the API interface between
infra and network. It is modelled as `Network` trait that provides two methods:

* `sink` - returns the sending side of the channel. All messages directed towards
the other helper parties must be written into it.`Sink` interface is a quite
power abstraction with flow control and flush mechanisms built-in. There is no
use of them yet, but we may decide to just buffer everything inside the `Sink`.
* `stream` - the incoming flow of messages sent by other helpers.

Note that it is possible and maybe even desirable to require callsites to provide
`channel_id` argument to these methods. That would mean keeping a logical "channel"
interface between infra and network. This would complicate the implementation of
both layers, but may be deemed necessary to implement as it better represents the
real world of IPA where multiple channels are used at the same time and it is preferred
to avoid contention between different channels.