Remove identity method from `Mesh`
Given the conversation in #110 [comment](https://github.com/martinthomson/raw-ipa/pull/110/files#r992706026) we don't seem to need to dynamically change the identities of helpers dynamically. Therefore we don't need to force client code to obtain a reference to the channel first.

The standard way to access helper identity should be via `ProtocolContext` and, in some cases, via `Gateway`