`Mesh` methods take shared reference
There is currently no need to take an exclusive reference to self. Given that `Mesh` is no longer a trait, this change is fairly easy to make.

Taking a shared reference instead of `&mut self` enables protocols to run multiple concurrent multiplications at the same time.

The use case for it is presented in #128