HTTP2/HTTP 1.1 metrics for IPA server
This change makes IPA web server to emit a metric that shows which
HTTP protocol was used for every request made to the helper. The goal
is to be aware if helpers start using HTTP 1.1. Restricting server code
to accept HTTP 2.0 only is not great because HTTP 1.1 is way easier to debug
and therefore valuable to support.

Alternatively we could just use a config that instructs the server which
protocol it should exclusively support. Right now it feels that having this
visibility is enough, but we may reconsider this in the future.

Note that it may be a bit misleading if connection starts as HTTP/1.1 and then
gets upgraded to HTTP/2.0 but in our setting helper clients should probably
initiate HTTP/2 from the beginning.