Byte serialization for field values
As I am preparing to start using contiguous buffers on infra side, serde serialization of messages and fields needs to get out of the way. Fields need to be able to transform themselves into a sequence of bytes and back because that is how infra will store and interpret them.

This also removes the annoying constraint that field values need to be wrapped before protocols can send them, so this change takes care of that too.

This was previously part of #150 but it grew too large and no longer reviewable in my opinion. So I am submitting it piece by piece.