Validate indices used in PRSS
Similarly to validating steps while narrowing contexts, this change makes PRSS validate the indices passed on to it when generating randomness.

It should be a mistake to use the same index twice within the same PRSS instance. This change enables PRSS to panic and abort in such cases.

Interestingly it already helped to catch one bug inside the malicious implementation.