Remove deny(warnings)
This setting prevents compiler from building the code with warnings.
I think it is useful to be able to do that because often I prototype
an implementation and want to see whether it compiles and it is ok
if there are unadressed warnings such asunused imports.
Also I want to be able to run tests and get feedback from them without
addressing all the warnings.

I think we have good protection at PR level as GH action
would prevent pushing code that violates this no warnings rule.