Run lint/Clippy first before build in GH actions
Most of the time, PR is broken because of formatting and/or Clippy issues. Therefore, it is desirable to run these checks first and get the feedback earlier before running builds and tests. My observation (likely biased towards my own behavior) is that we rarely submit code that does not compile or tests that are broken.