Make FSV a little leaner
This doesn't need to hold the batch size, it can be told that as an
argument to take().

It doesn't need to return what was previously there when inserting
values: that is a bug and we can assert if it happens.

Rename some variables and functions to use shorter names.

(I'm still not a fan of this taking relative offsets: it makes the
encapsulation incomplete.  That can wait though.)