Add shuttle and randomized test for sort
Unfortunately I can't enable this test right now because we have
a bug somewhere. Using this randomized scheduler revealed a bug in reveal (!?).
Reveal reconstructs wrong secret for some of the permutation values. I am not
sure where the bug is, likely in infra but it is certainly triggered by reordering
task execution. Will be looking into that, submitting this PR if someone else
wants to take a look  too.

Here is how to run this test

```bash
cargo test --features shuttle --lib tests::randomized::sort
```

I also refactored the sort circuit so we can run the same test in benchmarks, unit tests
and randomized tests