Introduce concurrency testing framework
This is mostly a mechanical change to set up IPA code to support Shuttle framework. The main change is to replace std imports with shuttle if `shuttle` feature is enabled. Unfortunately that means that unit tests need to use additional flag that disables them if `shuttle` is enabled.

There are two new concurrency tests added - they do not pass and require a fix in Infra. I just want to make sure that change gets properly reviewed and it is not hidden behind a pile of refactoring. I will be sending the fix after.