Enable test world to print metrics per step
Brought `metrics-runtime` crate because it supports scoped sinks, so step strings can be shared and stored once. Moreover it provides a nice hierarchy where `Gateway` for example does not need to know the helper role. All metrics emitted from the `Gateway` will be prefixed with the role.

There was quite a bit of plumbing to get the sink inside `IndexedEndpoint`, it looks a bit suboptimal to do it this way.

To force tests to print these metrics `RUST_LOG=raw_ipa::test_fixture::metrics=TRACE` must be set

### Sample output

```bash
2022-11-29T08:00:19.357701Z TRACE raw_ipa::test_fixture::metrics: Helper.Step,Records sent,Indexed PRSS,Sequential PRSS
2022-11-29T08:00:19.357760Z TRACE raw_ipa::test_fixture::metrics: H1.protocol/run-0/sort0/bit_permute,100,100,0
2022-11-29T08:00:19.357773Z TRACE raw_ipa::test_fixture::metrics: H1.protocol/run-0/sort0/mod_conv/xor1,50,50,0
2022-11-29T08:00:19.357781Z TRACE raw_ipa::test_fixture::metrics: H1.protocol/run-0/sort0/mod_conv/xor2,0,50,0
```