Improve error message for duplicates
Before this change, infrastructure did not allow duplicates to be inserted (duplicate is a message with the same record id sent to the same channel more than once). That is expected but the error did not get propagated back to the caller (protocol in this case). `send` method did not wait for the status of the operation - regardless whether gateway can actually accept this message or no, it allowed protocols to move forward immediately after send.

New behavior provides the essential mechanism to built the backpressure later, but currently it is used to propagate errors from the send buffers.

Protocols that attempt to insert the same message twice to the infrastructure will get an error at the point where this insertion occurs and will get the proper message

```
An error occurred while sending data to channel[peer=H2,step=step=protocol/run-0]: Record RecordId(1) has been received before
```