Print helper state if nothing is going on
It is an easy mistake to make to start counting records from 1 or miss sending a record. Infrastructure stays silent and does not output anything for it. This change sets up a timer for gateway loop that prints the state of helper (if there is something pending) every 10 seconds. The output looks like this

```bash
2022-12-05T05:48:59.379513Z ERROR raw_ipa::helpers::messaging: Nothing has happened on H1 in the last 10s. Here is the list of tasks pending completion:
 waiting to send: [
channel[peer=H2,step=step=protocol/run-0]: [0]
],
 waiting to receive: [
channel[peer=H2,step=step=protocol/run-0]: [1, 2, 3, 4, 5, 6, 7, 8, 9]
]
```