Gateway: do not report buffers that are 100% empty
It causes noise when we have long running tests. The proper solution
would be to implement channel closure https://github.com/private-attribution/ipa/issues/168
as it can be seen how much memory we're wasting by not doing it.

Skipping empty buffers is correct because there is no error on the protocol side