A better arithmetic circuit setup
I found that this file was completely unusable, so I fixed it and it
helped me understand that perhaps we have a bit of a problem to address
with our buffering.

Test this out by running:

```
cargo bench -F test-fixture --bench oneshot_arithmetic -- --width 1 --depth 2
```

Then try again after setting `config.gateway_config.send_buffer_config.items_in_batch = 2`.

In this test we construct a very simple circuit where we multiple 1 by 1
several times in a chain, like so:

```
let a = share(1);
for i in 0..depth {
  let b_i = share(1);
  a = a * b_i; // multiply with record i
}
return a;
```

This works perfectly when the buffer is full, but it fails when the
output of the multiplication with record 1 depends on the output of the
multiplication with record 2.  Because the first multiplication won't
complete until communication occurs and we don't communicate until we
have two records.

I don't know how often we iterate through records and await the outcome
of a multiplication on record i before moving on to record i+1, but it
seems like that might be fairly common.  If it is, then we have a pretty
serious problem.

[ci skip]