Experimental serialize trait based off GenericArray
`serialize`/`deserialize` now work with fixed-size arrays. Unfortunately it does not eliminate all the runtime checks because in some places I had to use `GenericArray::from_slice`.

The benefit of this approach is that `Serialize` implementors do not need to check the size of the buffer, it is guaranteed to be set correctly.