Remove trailing bits from the bit decomposition protocol
We just throw these out, so remove 'em.  The function now operates
modulo 2^l, which is all we use.  We can't do the same for the earlier
addition as we need to know about the overflow.

Savings are modest.  So 3 multiplications per event.
27 records in semi-honest; 54 in malicious for our 9 event test.

I also renamed both functions.  The old names annoyed me, because context
provided a lot of information that was again repeated in the name.  The
new names are shorter.  We should remove "bitwise" from other "boolean"
functions, because bitwise is implied in most cases.