A test coverage script
This is a little bit of an indulgence for me, but I find that generating
coverage reports was a little bit tedious to setup.  This script can be
pointed at a file to generate a coverage report.  It's a little
elaborate, but it automates the process completely, so it might be
useful.

I haven't enabled cargo options (like `-F`) but those can be added.