A substitute for join_all()
This is much like `StreamExt::buffered`, but it has further restrictions
on its use that might make it more performant.  The underlying
`Buffered` implementation there uses `FuturesOrdered`, which polls in an
arbitrary order.  Doing that ensures that progress can be made without
depending on properties of the underlying futures, but it means that
`FuturesOrdered` needs to be quite a bit more careful in its
construction, leaving some optimizations unrealized.

This implementation uses a simple `VecDeque` to back it, rather than a
linked list.  This makes it simpler, but less flexible.  If a future
resolves out of order, this panics.  That's very much not a safe way to
build something generic, but if you have some assurance that the futures
you provide will always resolve in order (which `UnorderedReceiver` and
`OrderingSender` can ensure), then it greatly simplifies the
implementation.

This isn't really finished.  The next version of this will take a stream
as input, which will be far more flexible.  After that, there is a need
to implement a `try_` version for when the futures return `Result`.