create separate server / network config structs
`NetworkConfig` contains information describing the helper network
as a whole. The intent is that this can be the same on all three
servers. `ServerConfig` is for a single server and has information
unique to the local server (e.g. where to find private keys).

Fix use of kernel-selected free ports for tests (#508).

Removed specification of identities when creating a `Processor`. The
association of helper identites 1/2/3 with discovery information is
established in `NetworkConfig`.