Use two channels inside random bits generator
This is just a prelude to a larger refactoring after which `RandomBitsGenerator` will be able to efficiently use the infrastructure buffering capabilities.

Currently, it does not change anything for the callers, but the idea is that `fallback_generator` will opt out from using buffering and will be continuously sending data as it goes in. With the right field size, the probability of `RandomBitsGenerator` calling it more than once is negligible, calling it once for `Fp32BitPrime` is $10^-9$