Fix send gateway buffer capacity
As Martin predicted, send capacity was destined to be broken because
of the bad config name. So we ended up with send buffer capacity that is
4 times smaller than we wanted.

I thought it might be a good idea to align buffer capacity with the field size
as it should work for the majority of the channels. That required making sure
IPA uses the same field, so I had to plug in a generic argument to it