Add bytes.sent metric to count bytes
A companion to `records.sent` that measures the total communication
in bytes between helpers. With recent developments in IPA, the total
number of records may stay the same, but size of messages sent is reduced.
`records.sent` does not capture it, so we would rely on `bytes.sent` to check
those improvements.