Match the behavior of IPA in the clear and MPC IPA
Found this issue while working on some CLI tooling. Currently MPC
just skips the attribution window logic if the window is set to be 0.

IPA in the clear did not have it and I was observing different results there.

It does seem better to use the type system to make this behavior more clear.
As 0 truly does not make much sense as window value, `Option<NonZeroU32>>` is
used instead.