Make helpers capable of running more than one query
This change allows both in memory and HTTP implementations to run more than one query in sequence.
There were two things preventing it: `QueryProcessor` not properly cleaning up its state and transports keeping the stream
collections after query completion. The former was easy to fix, but I am not super happy with the fix for the latter.

There were a couple of options how to fix the transports:

* Generate a unique query id for each run (#615). We need this anyway, so this could be an incomplete fix.
  Incomplete because one may consider keeping streams for completed query around forever as memleak.
* Teach `Transport` trait to be aware of query completion. That looked as too big of a fix and a violation of the design we have right now
* Implement `reset` for both transports that cleans up their internal state.
  That's the approach taken in this change. HTTP transport is aware of query completion callback,
  so changing it was easy, in memory transport does not directly interact with it, so the fix is implemented at `TestApp` level