Reduce the size of Docker build context
Before this change, we had Docker transferring gigabytes (if both release and debug builds exist) of data inside the image and make no use of it. It was happening because docker [expects](https://docs.docker.com/engine/reference/builder/#dockerignore-file) `.dockerignore` file to be inside the root folder of the context.

In IPA folder structure, `.dockerignore` was put right next to the `.Dockerfile` so it wasn't picked up when building the image.

There does not seem to be an option to provide path to `.dockerignore` to Docker CLI, so I build the context manually