Remove the limit on number of concurrent streams
See #685 for details. We certainly need more than 256 streams. Having an upper limit close to that number leaves us vulnerable to having this issue again in the future.

So I'd rather not limit it and rely on performance metrics if we start seeing impact of too many streams open.

I also don't see an easy way to test this, so tests are not included :(