Brush up hpke module
As we are preparing to create a separate aggregation circuit, we need better names for HPKE primitives used in IPA. We no longer encrypt shares of match keys only, for aggregation we want to encrypt boolean shares of various sizes. We will start with 1 byte shares

This change just moves things around and does a bunch of renames. No functional changes