Goodness gracious, it compiles!
If this plane did not have a WiFi, I'd never be able to do it.

It seems that `Neg` was tripping me off - adding it as a constraint to trait <A, B> and only constraining it to B, makes the compiler go to infinite loop trying to evaluate all possible types that can fit. At the end, I was getting the gnarly "overflow evaluating ..." error

So structure needs to be as follows:

* Traits that have Rhs and Output
* Traits that only have Output
* Traits that are valid for secret sharing (add, sub, neg)
* Traits that are valid for basic arithmetic (mul)