Rename WeakSharedValue
Long ago, `SharedValue` was a base trait common to `Field` and
`BitArray`. At some point, `BitArray` became `GaloisField` and was
changed to require `Field`. At that point, I think `SharedValue` and
`Field` became effectively synonymous.

Recently, the need arose for SharedValues that do not have
multiplicative properties. This was introduced as a new
`WeakSharedValue`.

This change merges the notion of `SharedValue` into `Field` and then renames
`WeakSharedValue` to `SharedValue`.