A draft API for validation of replicated shares
This is a simple stream adapter that performs inline validation of
anything that is replicated.  It will error at the end of the stream if
it finds a mismatch.  (Future work might include periodic checkpoints.)

The reason I'm putting this up is that it is a great example of how our
borrowed contexts are a giant pain in the posterior.  I'm starting to
think that we need to think about moving to something with Arc so we can
avoid this lifetime business.  It's a real drag.