TestWorld for sharded environments
This change introduces the ability to run very simple circuits on multiple
shards in parallel. It is not possible to communicate between shards yet,
but it is possible to use the same test infrastructure to create multiple shards
and use PRSS inside them as well as provide the input for each shard and consume
their output.

The next and hopefully final change will bring the ability to communicate across
shards.