Use `BytesStream` inside `Transport` trait
This change is to unify stream handling between in-memory and real-world implementations. Before, in memory streams yielded `Vec<u8>` and HTTP streams operated on `Result<Bytes, _>`.

This discrepancy required us to build a compatibility layer, and it makes sense now to unify both implementations.

The main motivator for this change is to be able to use `RecordsStream` inside shard receivers - `UnorderedReceiver` use case fits MPC model well, but is clunky to use inside shard channels.