Enable dual-mode for `RecordsStream` (#1002)
* Enable dual-mode for `RecordsStream`

`RecordsStream` is a convenient abstraction that takes care of polling the underlying raw byte stream, buffering enough data and deserializing them into a client-provided type `T`. It also handles padding and unaligned errors.

Before this change, stream element could only be a `Vec<T>` which means an extra allocation for clients that want to handle items one by one. This is the case for shard receiving streams.

Shard receiving streams are polled by MPC protocols that get records out of them one-by-one. The key difference between MPC receiving streams and shard receiving streams is that the former wants the ability to receive elements out of order. For shard streams, it is not required.

Shard streams can benefit from using `RecordsStream` abstraction quite a lot. I have it in my private branch and it works well. This change enables `RecordsStream` to be used in mode where they yield only one element