Remove field type parameter from `BooleanProtocols`
This parameter is not completely redundant, in that boolean protocols
can be implemented in any prime field. But we only currently use
them with `Boolean`.

Our support for other fields is not complete, because we only implement
`Not` for `Boolean` (and arrays thereof). The most straightforward way
to implement `Not` for other fields is to compute `1 - x`, which
requires `ShareKnownValue`, which requires a context. But maybe we could
define a constant THREE_INV with the value 1/3 (as long as we don't need to
support $\mathbb{F}_3$) and implement context-less `Not` in other
fields.