Simplify malicious share upgrades, and fix collect-steps
An early version of the compact gate (see #711) required that the semi-honest
share upgrade narrow the context with a step name that was distinguishable from
the malicious version, so that it could be named elsewhere as a special case.
This is no longer required, so the malicious and semi-honest impls can be moved
back to default impls in the trait definition.

This change also fixes steps.txt generation, which broke due to additional
logging that is now output by the IPA runs.