Improvment to the DZKP Batch Storage (#1050)
The PR removes the option from vec, and changes some of the more complex and expensive functionalities.

It changes the seq_join function on the validator (now validated_seq_join) and separates the chunk_size parameter in two separate parameters: chunk_size is the amount of futures processed in a single chunk (i.e. DZKP Batch) and multiplication_amount which determines how much memory is allocated within a Batch.

chunk_size is now an input to validated_seq_join whereas multiplication_amount is specified at the time the validator is generated.

The PR does not make changes to the DZKP conversion which is a separate PR which I will also adjust as discussed: #1032

Upon request by Ben Savage and Alex Koshelev, the PR deletes the reallocation ability of the Batch. If the parameter multiplication_amount is set incorrectly, it will cause a panic and the program crashes (whereas previously it would recover). The benefit is a reduction in code complexity and better maintainability.