Use a dedicated step for PRSS
Test multiplication does not work properly with the previous implementation of compact gate. W/o narrowing to PRSS, we send one message for key exchange and then try to multiply starting from record_id = 0 which causes infra to complain.

It is not great that PRSS uses someone else's capacity to exchange messages, so this commit fixes that